all filters accepts props
