+++ frontend/src/components/results/filters/Actions.jsx
@@ -1,20 +1,17 @@
 import { Checkbox, Stack, Text} from '@chakra-ui/react'
-import { actions } from '@/components/data/FetchActionsData'
-
-const filteredActions = actions.filter(countSetToZero)
 
 function countSetToZero(action) {
   return action.count > 0;
 }
 
-console.log(filteredActions)
+export const Actions = ({title, checkboxArray}) => {
 
-export const Actions = (props) => {
+  const filteredActions = checkboxArray.filter(countSetToZero)
 
   return (
   <Stack spacing={[2]} direction={['column']}>
     <Text fontSize="lg" color="on-accent-muted" fontWeight="medium">
-      Actions
+      {title}
     </Text>
     {filteredActions.map((action) => (
         <Checkbox size='md' colorScheme='blue' key={action.action} id={action.action} value={action.action} defaultChecked>
