+++ frontend/src/components/results/filters/WeightSlider.jsx
@@ -1,7 +1,7 @@
 import { Box, RangeSlider, RangeSliderFilledTrack, RangeSliderMark, RangeSliderThumb, RangeSliderTrack, Text} from '@chakra-ui/react'
 import React, { useState } from 'react';
 
-function WeightSlider({ title, range, initial }) {
+export const WeightSlider = ({title, range, initial}) => {
   const [sliderValue, setSliderValue] = useState([25, 75])
   return (
     <>
@@ -31,6 +31,3 @@ function WeightSlider({ title, range, initial }) {
     </>
   )
 }
-
-
-export default WeightSlider;
