+++ frontend/src/pages/adverseEvents/index.js
@@ -4,7 +4,34 @@ import { Heading } from '@chakra-ui/react'
 import ResultsLayout from '@/components/results/ResultsLayout';
 import DataTable from '@/components/results/DataTable'
 import theme from '@/styles/theme';
-import { AETargetSidebar } from '@/components/results/AETargetSidebar';
+import { Actions } from '@/components/results/filters/Actions';
+import { WeightSlider } from '@/components/results/filters/WeightSlider';
+import { Descriptors } from '@/components/results/filters/Descriptors';
+import { actionsData } from '@/components/data/FetchActionsData';
+
+const checkboxData = [
+  {
+    name: 'Gene',
+  },
+  {
+    name: 'Protein'
+  },
+  {
+    name: 'GWAS'
+  },
+  {
+    name: 'Phenotype'
+  },
+  {
+    name: 'Reactome'
+  },
+  {
+    name: 'Signor'
+  },
+  {
+    name: 'IntAct'    
+  }
+]
 
 
 export default function TargetToAEResults() {
@@ -15,14 +42,30 @@ export default function TargetToAEResults() {
         </Head>
         <Box display='flex' w="100%">
       
+
+        {/* Results Page Filters */}
         <Box w="25%" minW='250px'>
-          <AETargetSidebar />
+          <Stack spacing={[4]} direction={['column']} p='5'>
+            <Heading fontSize='xl'>Filters</Heading>
+            <Divider />
+            <Descriptors 
+              title = 'Descriptors'
+              checkboxArray = {checkboxData} />
+            <Divider />
+            <Actions 
+              title = 'Actions'
+              checkboxArray = {actionsData} />
+            <Divider />
+            <WeightSlider
+              title='Adverse Event Name'
+              range={{min: '0', mid: '50', max: '100'}} 
+              initial={{ min: '25', max: '75'}}/>
+          </Stack>
         </Box>
         <Box p={5} w="75%" bg="#eee">
-
-          {/* Search Page Heading */}
+          {/* Result Page Heading */}
           <Heading size='md' mb={4}>Target to Adverse Event Results</Heading>
-          {/* Search Input Display */}
+          {/* Results Input Display */}
           <Box display='flex' alignItems='center' justifyContent='space-between' mb='5'>
             <Box w="50%" bg="#eee">  
             <Card>
@@ -43,7 +86,7 @@ export default function TargetToAEResults() {
             </Card>
             </Box>
           </Box>
-          {/* Search Results Table */}
+          {/* Results Table */}
           <Box w='100%' mb='5'>
           <Heading size='md' mb={4}>Adverse Events Associated with Target</Heading>
           <DataTable/>
